upd: move SpiderLoggerAdapter to utils.log
